refactor to simplify some of the rendering logic
